src/fin_chatbot/services/requirements.txt
arrow==1.3.0
binaryornot==0.4.4
blinker==1.8.2
certifi==2024.8.30
chardet==5.2.0
charset-normalizer==3.3.2
click==8.1.7
cookiecutter==2.6.0
Flask==3.0.3
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdurl==0.1.2
Pygments==2.18.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-slugify==8.0.4
PyYAML==6.0.2
requests==2.32.3
rich==13.8.1
six==1.16.0
text-unidecode==1.3
types-python-dateutil==2.9.0.20240906
urllib3==2.2.3
Werkzeug==3.0.4
yfinance==0.2.26  # Or the version you installed
redis==5.1.0
ratelimit==2.2.1
psycopg2-binary==2.9.9
finnhub-python==2.4.20
prometheus_client==0.21.0
scipy==1.13.1
fasteners==0.19

# Additional dependencies for Scrapy, LangChain, and OpenAI
scrapy==2.10.1
langchain==0.0.300
faiss-cpu==1.7.4
openai==0.27.8